requirements.txt
pandas
fastparquet
django
djangorestframework